+++ src/main/java/lotto/controller/controllers/LottoResultController.java
@@ -1,4 +1,4 @@
-package lotto.controller;
+package lotto.controller.controllers;
 
 import lotto.domain.WinningLotto;
 import lotto.dto.BuyLottoDTO;
